Adds script to update translation files (#26)
* Renames install script

Renames `install.sh` to `messagingmenu.sh` to better reflect the script's purpose as a general utility for the messaging menu extension.

Updates the README to reflect the script's name change and the new installation command.

* Adds translation update command

Introduces a new command to update translation files.

This replaces the separate shell script with a command
integrated into the main script. It simplifies the process of
updating translations.